embedded net exports

+++ experiments/car_detection_magnetic_field/models/net_3/export/MyModel.h
@@ -4,7 +4,7 @@
 
 #include <ModelInterface.h>
 
-class MyModel : public ModelInterface<float>
+class MyModel : public ModelInterface<int8_t>
 {
 	public:
 		 MyModel();
